Readme for firmware >= 3.3.9_003077 (Vacuum robot)
Hey, feel free to refactor it if you want to. It took me ages before I found that. It was working on the xiaomi wifi network but I was getting timeout error without any other explanation once the vacuum was connected to my local network...
Thanks for this project btw :)