Add last_clean_details to return information from the last clean
This involves also refactoring clean_details() to return the results unwrapped.
For the time being the default behavior will remain the same as it was (wrapped),
but a deprecation warning will be raised.

I'm expecting that we can flip the toggle after a while, and remove the warning
at that point, too.