Ignore .venv directory and docs _build directory in git
We don't want those directories commited to the git repository.

Python virtual environments are often created in a .venv/ directory,
thus ignoring all .venv/ directories.

The docs/_build directory is automatically created by Sphinx when the
documentation is generated. We don't want the generated documentation
to be commited to the repository.