Sphinx config fix (#458)
* Ignore .venv directory and docs _build directory in git

We don't want those directories commited to the git repository.

Python virtual environments are often created in a .venv/ directory,
thus ignoring all .venv/ directories.

The docs/_build directory is automatically created by Sphinx when the
documentation is generated. We don't want the generated documentation
to be commited to the repository.

* Ignore .venv directory in Sphinx config

This is required because if someone stores the virtualenv in a .venv
directory (common practice) under the docs/ path, then Sphinx will try
to pickup all files under the virtualenv. This leads to a lot of errors,
performance impacts and side effects.

* Remove static HTML path from Sphinx config

There are currently no additional static files for Sphinx in place,
but there's a _static directory configured in the Sphinx docs. When
creating the Sphinx docs without an `mkdocs _static` command before,
this leads to an error.

Thus removing the _static path until we need it again.