Add support for Xiaomi Aqara Gateway Alarm
I am working on integrating the Xiaomi Aqara Gateway Alarm into HomeAssistant as an official component.
Since the "Xiaomi Aqara API (LAN protocol)" does not support the Alarm of the Gateway I use the "Xiaomi's miIO protocol" as implemented in your awesom library.

I know that HomeAssistant requires commans such as included in this PR to be in the upstream library instead of directly in the HomeAssistant code. That is why I want this merged in python-miio.

@rytilahti Could you please merge this PR and push to a new version of python-miio such that I can then create the PR in HomeAssistant to add official support of the Xiaomi Gateway Alarm?

Thank you in advance for your time and effort!