Remove __json__ boilerplate code from all status containers
All implementations were simply returning the `data` for this request,
and forgetting to add this boilerplate piece causes problems like shown in #816.

This commit adds a lookup for `data` variable which should be consistent within all containers.
For the time being, this behavior can still be overridden by manually defining __json__.

Fixes #816