Add support for Dreame Vacuum MB1808 1C
The full robot API is implemented in a dictionary which is interpreted
by helper functions in DreameVacuumBase. There is also an example which
allows to navigate the robot by the keyboard arrows.
Even though the coding conventions of the project are not followed, this
contribution might be helpful for the project.