Drop support for python 3.7
I would prefer to drop support for python 3.8 too, as the debian oldstable
shipping 3.7 is going to be EOL'd soon and the current stable has 3.9.
However, pypistat says that it seems to be still a fairly widely, which
is surprising to me.

Any ideas why that is the case?

❯ pypistats python_minor python-miio
| category | percent | downloads |
|:---------|--------:|----------:|
| 3.9      |  61.29% |   199,700 |
| 3.10     |  18.12% |    59,039 |
| 3.8      |  16.87% |    54,956 |
| 3.7      |   2.15% |     7,010 |
| null     |   0.98% |     3,199 |
| 3.6      |   0.26% |       848 |
| 2.7      |   0.23% |       744 |
| 3.5      |   0.09% |       285 |
| 3.11     |   0.01% |        37 |
| 3.4      |   0.01% |        18 |
| Total    |         |   325,836 |

Date range: 2022-01-18 - 2022-07-17