Add introspectable settings
The @settings decorator creates a corresponding descriptor
depending on what parameters are given to it:
* If min_value or max_value is defined, a NumberSettingDescriptor is created.
* If choices is defined, an EnumSettingDescriptor is created.

This also converts the descriptors to use attrs as kw_args is only available from python 3.10 onwards,
and using it is a prerequisite to allow sane inheritance on data classes.

This commit also touches cooker.py to rename previously used variable `settings` to `_settings` and its exposed getter to `cooker_settings`.