Implement embedding DeviceStatus containers
Adds DeviceStatus.embed(self, other: DeviceStatus) method that allows
embedding separate DeviceStatus containers into a single one.
This will copy over the switch, setting, and sensor descriptors to the
container and prefixing their names with the name of the other class.

Internally, __getattribute__ is overridden to check if the property name
contains ':' and if so, the attribute lookup is done in an embedded class
named after the first part of the key.